requirements.txt
pandas
numpy
scikit-learn
seaborn
mlflow
python-dotenv
pymongo
certifi
pymongo[srv]==3.12
pymongo
dagshub
fastapi
uvicorn
python-multipart

# -e .